Logout Fn and Error Messages - moved the logout function back to the main.js because it looks cleaner there! - added utils.js for error message handling - updated some HTML files that will need error handling - updated SCSS file for error messages - Few minor tweaks to the header
