+++ css/main.css
@@ -6843,6 +6843,8 @@ footer {
   vertical-align: sub; }
 
 .hidden {
+  height: 0;
   opacity: 0;
+  padding: 0;
   transition: all 0.3s ease;
   visibility: none; }
