+++ index.html
@@ -22,6 +22,7 @@
 
         <!-- start of form template -->
         <form id="form" class="form box">
+          <p class="notification hidden has-text-danger">Password does not match!</p>
           <span class="columns">
             <label class="label column is-full" for="email">Email:
               <input class="input" id="email" type="email" placeholder="email@address.com" required validate>
@@ -37,7 +38,6 @@
             <button class="button is-primary" type="submit">Login</button>
             <a class="tag has-text-link" href="./signup.html">Create an Account</a>
           </span>
-          <p class="notification has-text-danger">Password does not match!</p>
         </form>
         <!-- end of form template -->
