+++ scripts/utils.js
@@ -1,0 +1,8 @@
+const errorMessage = (container, message, time) => {
+  const notice = document.querySelector(container)
+  notice.innerHTML = message
+  notice.classList.remove('hidden')
+  setTimeout(() => { notice.classList.add('hidden') }, time)
+}
+
+module.exports = { errorMessage }
