+++ signup.html
@@ -20,6 +20,7 @@
 
         <!-- start of form template -->
         <form id="form" class="form box">
+          <p class="notification hidden has-text-danger">Email account already in use!</p>
           <span class="columns">
             <label class="label column is-half" for="first_name">First Name:
               <input class="input" id="first_name" type="text" placeholder="First Name" required validate>
@@ -50,7 +51,6 @@
             <input class="button is-primary" type="submit" value="Create an Account">
             <a class="tag has-text-link" href="./index.html">Already have an account?</a>
           </span>
-          <p class="notification has-text-danger">Email account already in use!</p>
         </form>
         <!-- end of form template -->
